Flattened compents for the workout screen by killing off set.js and updated refrences as such
